tests/wpt/tests/svg/animations/svglength-animation-values.html

<!doctype html>
<html>
<meta charset="utf-8">
<title>Test SVGLength animation set with 'values', different units and different count of spaces.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup test document
var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("x", "0");
rect.setAttribute("width", "100");
rect.setAttribute("height", "100");
rect.setAttribute("fill", "green");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animate");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "width");
animate.setAttribute("begin", "0s");
animate.setAttribute("dur", "4s");
animate.setAttribute("values", "100px;  130  ;4cm  ;6in; 200pt");
rect.appendChild(animate);
rootSVGElement.appendChild(rect);

// Setup animation test
function sample1() {
    // Check initial/end conditions
    assert_approx_equals(rect.width.animVal.value, 100, epsilon);
    assert_equals(rect.width.baseVal.value, 100);
}

function sample2() {
    assert_approx_equals(rect.width.animVal.value, 130, epsilon);
    assert_equals(rect.width.baseVal.value, 100);
}

function sample3() {
    assert_approx_equals(rect.width.animVal.value, 151.2, epsilon);
    assert_equals(rect.width.baseVal.value, 100);
}

function sample4() {
    assert_approx_equals(rect.width.animVal.value, 576, epsilon);
    assert_equals(rect.width.baseVal.value, 100);
}

function sample5() {
    assert_approx_equals(rect.width.animVal.value, 267, epsilon);
    assert_equals(rect.width.baseVal.value, 100);
}

smil_async_test((t) => {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 0.0,   sample1],
        ["animation", 1.0,   sample2],
        ["animation", 2.0,   sample3],
        ["animation", 3.0,   sample4],
        ["animation", 3.999, sample5],
        ["animation", 4.001, sample1]
    ];

    runAnimationTest(t, expectedValues);
});

</script>
